Alternate (fixed) version of images in lists
The images are now loaded correctly in the lists.
Images are cached in an LRU fashion in memory with a max
size based on device memory class.

Also made EventDetailActivity use the same image loader and
cache so we don't have to reload those images from the server
so often.

This could of course be perfected by adding a file cache later.